requirements.txt
Django===3.2.16
xlrd===2.0.1
python-docx===0.8.11